refactor : LineTest, SectionsTest 분리 및 추가
SectionsTest: 노선 관리 테스트
LineTest : 노선 고유의 테스트
